+++ samples/shared/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/ui/test/TempAndroidTestScreen.kt
@@ -10,6 +10,7 @@ import com.bumptech.glide.load.DecodeFormat
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.github.panpf.zoomimage.GlideZoomAsyncImage
 import com.github.panpf.zoomimage.compose.glide.ExperimentalGlideComposeApi
+import com.github.panpf.zoomimage.compose.zoom.ScrollBarSpec
 import com.github.panpf.zoomimage.images.ComposeResImageFiles
 import com.github.panpf.zoomimage.sample.ui.base.BaseScreen
 import com.github.panpf.zoomimage.sample.ui.base.ToolbarScaffold
@@ -31,6 +32,7 @@ fun GlideThumbnailSample() {
         model = ComposeResImageFiles.hugeChina.uri,
         modifier = Modifier.fillMaxSize(),
         contentDescription = null,
+        scrollBar = ScrollBarSpec.DefaultAndWindowInsets,
         requestBuilderTransform = {
             it.thumbnail(
                 Glide.with(context)