+++ zoomimage-compose-coil3-core/src/commonMain/kotlin/com/github/panpf/zoomimage/CoilZoomAsyncImage.kt
@@ -19,6 +19,9 @@ package com.github.panpf.zoomimage
 
 import androidx.annotation.MainThread
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.safeDrawing
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.NonRestartableComposable
@@ -314,6 +317,10 @@ private fun CoilZoomAsyncImage(
             Box(
                 Modifier
                     .matchParentSize()
+                    .let {
+                        if (scrollBar.enabledWindowInsets)
+                            it.windowInsetsPadding(WindowInsets.safeDrawing) else it
+                    }
                     .zoomScrollBar(zoomState.zoomable, scrollBar)
             )
         }